// packages/ml-sdk/src/generated/shared/utils/common-types.ts
/**
*
*    Generated file, please do not change!!!
*    From http://www.vrap.io/ with love
*
*                ,d88b.d88b,
*                88888888888
*                `Y8888888Y'
*                  `Y888Y'
*                    `Y'
*
*/


export type MethodType =
  | "GET"
  | "HEAD"
  | "POST"
  | "PUT"
  | "DELETE"
  | "CONNECT"
  | "OPTIONS"
  | "TRACE";

export type QueryParam =
  | string
  | string[]
  | number
  | number[]
  | boolean
  | boolean[]
  | undefined;

export type VariableMap = {
  [key: string]: QueryParam;
};

export interface ClientRequest {
  baseUri?: string;
  uri?: string,
  headers?: VariableMap;
  method: MethodType;
  uriTemplate?: string;
  pathVariables?: VariableMap;
  queryParams?: VariableMap;
  body?: any
}

export type ClientResponse<T = any> = {
  body: T;
  statusCode?: number;
  headers?: Object;
};

export type executeRequest = (request: ClientRequest) => Promise<ClientResponse>